Move summary sorting and enumeration to EventScoreViewModel and use configured classes
